backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
numpy==1.24.3
opencv-python==4.8.1.78
face-recognition==1.3.0
mediapipe==0.10.8
pydantic==2.5.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
eyetrax==0.2.2
pymongo==4.6.0
google-auth==2.25.2
cryptography==41.0.7
passlib[bcrypt]==1.7.4

